requirements-cloud.txt
# Streamlit Cloud Deployment Requirements
# Includes pysqlite3-binary for Linux/Cloud deployment

streamlit>=1.30.0
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-cohere>=0.1.0
langchain-community>=0.0.13
chromadb>=0.4.18
pysqlite3-binary>=0.5.0
cohere>=5.0.0
huggingface-hub>=0.20.0
sentence-transformers>=2.3.0
pypdf>=4.0.0
python-dotenv>=1.0.0
